Move FaunadbClient inside sqlalchemy_fauna module
Since we only use the client via the SQLAlchemy Dialect now,
it makes sense to include it in the sqlalchemy module. I also
deleted some no-longer-used references to the client.
